Conflicts:
	examples/ucp_client_server.c
	src/ucp/api/ucp.h
	src/ucp/api/ucp_def.h
	src/ucp/core/ucp_am.c

+++ test/gtest/common/test_helpers.h
@@ -405,7 +405,7 @@ static void fill_random(C& c, size_t size) {
 template <typename T>
 static inline T random_upper() {
   return static_cast<T>((rand() / static_cast<double>(RAND_MAX)) *
-                        std::numeric_limits<T>::max());
+                        static_cast<double>(std::numeric_limits<T>::max()));
 }
 
 template <typename T>
